Consolidate C sources into one folder
