+++ .gitmodules
@@ -1,3 +1,3 @@
 [submodule "osqp"]
-	path = osqp_sources
+	path = c_sources/osqp_sources
 	url = https://github.com/osqp/osqp
